server/requirements.txt

fastapi
uvicorn
pydantic[email]
python-dotenv
python-multipart
bcrypt
pymongo
motor
minio
PyJWT
email-validator
pytest
pytest-mock
pytest-asyncio
